requirement-old.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bertviz==1.4.0
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==1.2.0
farasapy==0.0.14
filelock==3.12.0
Flask==2.2.5
fonttools==4.39.4
fsspec==2023.5.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
llvmlite==0.40.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
plotly==5.14.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
PyArabic==0.6.15
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
sklearn==0.0.post5
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0